add organizations list context
