+++ frontend/src/components/OrganizationsList/organizationsTable.scss
@@ -1,4 +1,4 @@
-.home {
+.organizations-table {
   tbody tr {
     &:hover {
       cursor: pointer;
