+++ frontend/src/contexts/index.ts
@@ -1,1 +1,2 @@
 export * from './UserContext';
+export * from './OrganizationsListContext';
